fix swig for python 3.9

+++ .github/workflows/install_deps.sh
@@ -45,7 +45,9 @@ do
       # PEtab
       # AMICI dependencies
       sudo apt-get install swig3.0 libatlas-base-dev libhdf5-serial-dev
-      sudo ln -s /usr/bin/swig3.0 /usr/bin/swig
+      if [ ! -e /usr/bin/swig ]; then
+        sudo ln -s /usr/bin/swig3.0 /usr/bin/swig
+      fi
     ;;
 
     doc)
